Added Global Packs support for Fabric also staged wrong file previously.
