+++ Common/src/main/resources/META-INF/services/subaraki.paintings.compat_layer.IPackRepoDiscoveryService
