+++ Fabric/build.gradle
@@ -34,6 +34,7 @@ dependencies {
     implementation project(":Common")
     modImplementation "${compat_mod}${compat_fabric}"
 
+    compileOnly group: 'com.electronwill.night-config', name: 'core', version: '3.6.5'//Used for Global Packs compat, Global Packs ships this
 }
 
 loom {
@@ -99,17 +100,17 @@ publishing {
     }
 }
 
-//task publishCurseForge(type: TaskPublishCurseForge) {
-//
-//    apiToken = "${api_token_paintings}"
-//
-//    // The main file to upload
-//    def mainFile = upload(252042, file("${project.buildDir}/libs/${archivesBaseName}-${version}.jar"))
-//    mainFile.releaseType = 'beta'
-//    mainFile.changelog = "${changelog}"
-//    mainFile.changelogType = 'markdown'
-//    mainFile.addRequirement('transparent')
-//    mainFile.addRequirement('cloth-config')
-//    mainFile.addRequirement('fabric-api')
-//
-//}
+task publishCurseForge(type: TaskPublishCurseForge) {
+
+    apiToken = "${api_token_paintings}"
+
+    // The main file to upload
+    def mainFile = upload(252042, file("${project.buildDir}/libs/${archivesBaseName}-${version}.jar"))
+    mainFile.releaseType = 'beta'
+    mainFile.changelog = "${changelog}"
+    mainFile.changelogType = 'markdown'
+    mainFile.addRequirement('transparent')
+    mainFile.addRequirement('cloth-config')
+    mainFile.addRequirement('fabric-api')
+
+}
