--- Fabric/src/main/resources/META-INF/services/subaraki.paintings.compat_layer.IPackRepoDiscoveryService
@@ -1,2 +1,0 @@
-subaraki.paintings.compat_layer.OpenLoaderCompat #OpenLoader
-#TODO Add one for Global Packs
